requirements.txt
telethon>=1.39.0
python-telegram-bot>=22.0
python-dotenv>=1.1.0
openai>=1.69.0
schedule>=1.2.0 